Clean up webserver handling
- Remove unused code
- Organize some parts better
